requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
transformers>=4.52.0
torch>=2.3.0
torchvision>=0.18.0
accelerate>=0.34.0
pillow>=10.4.0
opencv-python-headless>=4.10.0.84
python-multipart>=0.0.9
qwen-vl-utils>=0.0.9
requests>=2.32.0
